Move `CredentialLifetime` struct into `bot` package
