+++ lib/tbot/config/config.go
@@ -224,8 +224,8 @@ type BotConfig struct {
 	// top of the existing configuration during `CheckAndSetDefaults()`.
 	JoinURI string `yaml:"join_uri,omitempty"`
 
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
-	Oneshot            bool               `yaml:"oneshot"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
+	Oneshot            bool                   `yaml:"oneshot"`
 	// FIPS instructs `tbot` to run in a mode designed to comply with FIPS
 	// regulations. This means the bot should:
 	// - Refuse to run if not compiled with boringcrypto
@@ -422,7 +422,7 @@ func (conf *BotConfig) CheckAndSetDefaults() error {
 	}
 
 	// Validate CertificateTTL and RenewalInterval options
-	var ttlErr SuboptimalCredentialTTLError
+	var ttlErr bot.SuboptimalCredentialTTLError
 	err := conf.CredentialLifetime.Validate(conf.Oneshot)
 	switch {
 	case errors.As(err, &ttlErr):
@@ -430,7 +430,7 @@ func (conf *BotConfig) CheckAndSetDefaults() error {
 		// just reject the configuration in a future release.
 		//
 		//nolint:sloglint // msg cannot be constant
-		log.WarnContext(context.TODO(), ttlErr.msg, ttlErr.LogLabels()...)
+		log.WarnContext(context.TODO(), ttlErr.Message(), ttlErr.LogLabels()...)
 	case err != nil:
 		return err
 	}
@@ -446,7 +446,7 @@ type ServiceConfig interface {
 	// GetCredentialLifetime returns the service's custom certificate TTL and
 	// RenewalInterval. It's used for validation purposes; services that do not
 	// support these options should return the zero value.
-	GetCredentialLifetime() CredentialLifetime
+	GetCredentialLifetime() bot.CredentialLifetime
 
 	// GetName returns the user-given name of the service, used for validation
 	// purposes.
@@ -753,102 +753,4 @@ func ReadConfig(reader io.ReadSeeker, manualMigration bool) (*BotConfig, error) 
 	default:
 		return nil, trace.BadParameter("unrecognized config version %q", version.Version)
 	}
-}
-
-// CredentialLifetime contains configuration for how long credentials will
-// last (TTL) and the frequency at which they'll be renewed (RenewalInterval).
-//
-// It's a member on the BotConfig and service/output config structs, marked with
-// the `inline` YAML tag so its fields become individual fields in the YAML
-// config format.
-type CredentialLifetime struct {
-	TTL             time.Duration `yaml:"credential_ttl,omitempty"`
-	RenewalInterval time.Duration `yaml:"renewal_interval,omitempty"`
-	// skipMaxTTLValidation is used by services that do not abide by standard
-	// teleport credential lifetime limits to override the check that the
-	// user specified TTL is less than the max TTL. For example, X509 SVIDs can
-	// be issued with a lifetime of up to 2 weeks.
-	skipMaxTTLValidation bool
-}
-
-// IsEmpty returns whether none of the fields is set (i.e. it is unconfigured).
-func (l CredentialLifetime) IsEmpty() bool {
-	// We don't care about this field being set when checking empty state.
-	l.skipMaxTTLValidation = false
-	return l == CredentialLifetime{}
-}
-
-// Validate checks whether the combination of the fields is valid.
-func (l CredentialLifetime) Validate(oneShot bool) error {
-	if l.IsEmpty() {
-		return nil
-	}
-
-	if l.TTL == 0 || l.RenewalInterval == 0 {
-		return trace.BadParameter("credential_ttl and renewal_interval must both be specified if either is")
-	}
-
-	if l.TTL < 0 {
-		return trace.BadParameter("credential_ttl must be positive")
-	}
-
-	if l.RenewalInterval < 0 {
-		return trace.BadParameter("renewal_interval must be positive")
-	}
-
-	if l.TTL < l.RenewalInterval && !oneShot {
-		return SuboptimalCredentialTTLError{
-			msg: "Credential TTL is shorter than the renewal interval. This is likely an invalid configuration. Increase the credential TTL or decrease the renewal interval",
-			details: map[string]any{
-				"ttl":      l.TTL,
-				"interval": l.RenewalInterval,
-			},
-		}
-	}
-
-	if !l.skipMaxTTLValidation && l.TTL > defaults.MaxRenewableCertTTL {
-		return SuboptimalCredentialTTLError{
-			msg: "Requested certificate TTL exceeds the maximum TTL allowed and will likely be reduced by the Teleport server",
-			details: map[string]any{
-				"requested_ttl": l.TTL,
-				"maximum_ttl":   defaults.MaxRenewableCertTTL,
-			},
-		}
-	}
-
-	return nil
-}
-
-// SuboptimalCredentialTTLError is returned from CredentialLifetime.Validate
-// when the user has set CredentialTTL to something unusual that we can work
-// around (e.g. if they exceed MaxRenewableCertTTL the server will reduce it)
-// rather than rejecting their configuration.
-//
-// In the future, these probably *should* be hard failures - but that would be
-// a breaking change.
-type SuboptimalCredentialTTLError struct {
-	msg     string
-	details map[string]any
-}
-
-// Error satisfies the error interface.
-func (e SuboptimalCredentialTTLError) Error() string {
-	if len(e.details) == 0 {
-		return e.msg
-	}
-	parts := make([]string, 0, len(e.details))
-	for k, v := range e.details {
-		parts = append(parts, fmt.Sprintf("%s=%v", k, v))
-	}
-	return fmt.Sprintf("%s (%s)", e.msg, strings.Join(parts, ", "))
-}
-
-// LogLabels returns the error's details as a slice that can be passed as the
-// varadic args parameter to log functions.
-func (e SuboptimalCredentialTTLError) LogLabels() []any {
-	labels := make([]any, 0, len(e.details)*2)
-	for k, v := range e.details {
-		labels = append(labels, k, v)
-	}
-	return labels
 }
