+++ lib/tbot/config/service_application_tunnel.go
@@ -25,6 +25,7 @@ import (
 	"github.com/gravitational/trace"
 	"gopkg.in/yaml.v3"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/internal/marshaling"
 )
 
@@ -52,7 +53,7 @@ type ApplicationTunnelService struct {
 
 	// CredentialLifetime contains configuration for how long credentials will
 	// last and the frequency at which they'll be renewed.
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
 
 	// Listener overrides "listen" and directly provides an opened listener to
 	// use.
@@ -95,6 +96,6 @@ func (s *ApplicationTunnelService) CheckAndSetDefaults() error {
 	return nil
 }
 
-func (o *ApplicationTunnelService) GetCredentialLifetime() CredentialLifetime {
+func (o *ApplicationTunnelService) GetCredentialLifetime() bot.CredentialLifetime {
 	return o.CredentialLifetime
 }
