+++ lib/tbot/config/service_client_credential.go
@@ -27,6 +27,7 @@ import (
 	"golang.org/x/crypto/ssh"
 
 	"github.com/gravitational/teleport/api/client"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/identity"
 	"github.com/gravitational/teleport/lib/tbot/internal/marshaling"
 )
@@ -158,6 +159,6 @@ func (o *UnstableClientCredentialOutput) Type() string {
 	return UnstableClientCredentialOutputType
 }
 
-func (o *UnstableClientCredentialOutput) GetCredentialLifetime() CredentialLifetime {
-	return CredentialLifetime{}
+func (o *UnstableClientCredentialOutput) GetCredentialLifetime() bot.CredentialLifetime {
+	return bot.CredentialLifetime{}
 }
