+++ lib/tbot/config/service_database.go
@@ -101,7 +101,7 @@ type DatabaseOutput struct {
 
 	// CredentialLifetime contains configuration for how long credentials will
 	// last and the frequency at which they'll be renewed.
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
 }
 
 // GetName returns the user-given name of the service, used for validation purposes.
@@ -209,7 +209,7 @@ func (o *DatabaseOutput) Type() string {
 	return DatabaseOutputType
 }
 
-func (o *DatabaseOutput) GetCredentialLifetime() CredentialLifetime {
+func (o *DatabaseOutput) GetCredentialLifetime() bot.CredentialLifetime {
 	return o.CredentialLifetime
 }
