+++ lib/tbot/config/service_example.go
@@ -22,6 +22,7 @@ import (
 	"github.com/gravitational/trace"
 	"gopkg.in/yaml.v3"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/internal/marshaling"
 )
 
@@ -67,6 +68,6 @@ func (s *ExampleService) CheckAndSetDefaults() error {
 	return nil
 }
 
-func (s *ExampleService) GetCredentialLifetime() CredentialLifetime {
-	return CredentialLifetime{}
+func (s *ExampleService) GetCredentialLifetime() bot.CredentialLifetime {
+	return bot.CredentialLifetime{}
 }
