+++ lib/tbot/config/service_kubernetes_v2.go
@@ -56,7 +56,7 @@ type KubernetesV2Output struct {
 
 	// CredentialLifetime contains configuration for how long credentials will
 	// last and the frequency at which they'll be renewed.
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
 }
 
 // GetName returns the user-given name of the service, used for validation purposes.
@@ -168,6 +168,6 @@ func (s *KubernetesSelector) UnmarshalYAML(value *yaml.Node) error {
 	return nil
 }
 
-func (o *KubernetesV2Output) GetCredentialLifetime() CredentialLifetime {
+func (o *KubernetesV2Output) GetCredentialLifetime() bot.CredentialLifetime {
 	return o.CredentialLifetime
 }
