+++ lib/tbot/config/service_ssh_host.go
@@ -63,7 +63,7 @@ type SSHHostOutput struct {
 
 	// CredentialLifetime contains configuration for how long credentials will
 	// last and the frequency at which they'll be renewed.
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
 }
 
 // GetName returns the user-given name of the service, used for validation purposes.
@@ -127,6 +127,6 @@ func (o *SSHHostOutput) Type() string {
 	return SSHHostOutputType
 }
 
-func (o *SSHHostOutput) GetCredentialLifetime() CredentialLifetime {
+func (o *SSHHostOutput) GetCredentialLifetime() bot.CredentialLifetime {
 	return o.CredentialLifetime
 }
