+++ lib/tbot/config/service_workload_identity_api_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/workloadidentity/workloadattest"
 )
 
@@ -46,7 +47,7 @@ func TestWorkloadIdentityAPIService_YAML(t *testing.T) {
 				Selector: WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
-				CredentialLifetime: CredentialLifetime{
+				CredentialLifetime: bot.CredentialLifetime{
 					TTL:             1 * time.Minute,
 					RenewalInterval: 30 * time.Second,
 				},
