+++ lib/tbot/config/service_workload_identity_aws_ra.go
@@ -197,6 +197,6 @@ func (o *WorkloadIdentityAWSRAService) GetDestination() bot.Destination {
 	return o.Destination
 }
 
-func (o *WorkloadIdentityAWSRAService) GetCredentialLifetime() CredentialLifetime {
-	return CredentialLifetime{}
+func (o *WorkloadIdentityAWSRAService) GetCredentialLifetime() bot.CredentialLifetime {
+	return bot.CredentialLifetime{}
 }
