+++ lib/tbot/config/service_workload_identity_jwt.go
@@ -47,7 +47,7 @@ type WorkloadIdentityJWTService struct {
 
 	// CredentialLifetime contains configuration for how long credentials will
 	// last and the frequency at which they'll be renewed.
-	CredentialLifetime CredentialLifetime `yaml:",inline"`
+	CredentialLifetime bot.CredentialLifetime `yaml:",inline"`
 }
 
 // GetName returns the user-given name of the service, used for validation purposes.
@@ -117,8 +117,8 @@ func (o *WorkloadIdentityJWTService) GetDestination() bot.Destination {
 	return o.Destination
 }
 
-func (o *WorkloadIdentityJWTService) GetCredentialLifetime() CredentialLifetime {
+func (o *WorkloadIdentityJWTService) GetCredentialLifetime() bot.CredentialLifetime {
 	lt := o.CredentialLifetime
-	lt.skipMaxTTLValidation = true
+	lt.SkipMaxTTLValidation = true
 	return lt
 }
