+++ lib/tbot/config/service_workload_identity_jwt_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/botfs"
 )
 
@@ -36,7 +37,7 @@ func TestWorkloadIdentityJWTService_YAML(t *testing.T) {
 					Name: "my-workload-identity",
 				},
 				Audiences: []string{"audience1", "audience2"},
-				CredentialLifetime: CredentialLifetime{
+				CredentialLifetime: bot.CredentialLifetime{
 					TTL:             time.Minute,
 					RenewalInterval: 30 * time.Second,
 				},
