+++ lib/tbot/config/service_workload_identity_x509_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/botfs"
 )
 
@@ -36,7 +37,7 @@ func TestWorkloadIdentityX509Service_YAML(t *testing.T) {
 					Name: "my-workload-identity",
 				},
 				IncludeFederatedTrustBundles: true,
-				CredentialLifetime: CredentialLifetime{
+				CredentialLifetime: bot.CredentialLifetime{
 					TTL:             1 * time.Minute,
 					RenewalInterval: 30 * time.Second,
 				},
